+++ src/main/java/org/jabref/gui/actions/StandardActions.java
@@ -173,8 +173,7 @@ public enum StandardActions implements Action {
     OPEN_FORUM(Localization.lang("Online help forum"), Localization.lang("Online help forum"), IconTheme.JabRefIcons.FORUM),
     ERROR_CONSOLE(Localization.lang("View event log"), Localization.lang("Display all error messages")),
     SEARCH_FOR_UPDATES(Localization.lang("Check for updates")),
-    ABOUT(Localization.lang("About JabRef"), Localization.lang("About JabRef")),
-    SELECT_ICONS(Localization.lang("Open icon panel"), IconTheme.JabRefIcons.SELECT_ICONS);
+    ABOUT(Localization.lang("About JabRef"), Localization.lang("About JabRef"));
 
     private final String text;
     private final String description;